refactor: migrate to MySQL and consolidate DbContext for URL shortener
