--- src/Backend/URLShortenerBackend/Entities/Url.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace URLShortenerBackend.Entities;
-
-public partial class Url
-{
-    public int Id { get; set; }
-
-    public string OriginalUrl { get; set; } = null!;
-
-    public string ShortCode { get; set; } = null!;
-
-    public DateTime CreatedAt { get; set; }
-
-    public DateTime ExpiresAt { get; set; }
-}
