+++ src/Backend/URLShortenerBackend/Models/ShortUrl.cs
@@ -3,12 +3,13 @@ namespace URLShortenerBackend.Models
     public class ShortUrl
     {
         public int Id { get; set; }
-        public string OriginalUrl { get; set; }
-        public string ShortCode { get; set; }
-        public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
+
+        public string OriginalUrl { get; set; } = null!;
+
+        public string ShortCode { get; set; } = null!;
+
+        public DateTime CreatedAt { get; set; }
 
         public DateTime? ExpiresAt { get; set; }
-
-
     }
 }
